Add Sidebar, fix Firebase config, update readme
This commit removes an annoying impediment which should never have
existed in the first place. For some reason I thought there was a
`firestore/provider` import necessary to get Google Auth working. This
was not necessary and is very wrong. There is no `firestore/provider`
NPM package. I should have known something was up with this hiccup in
development because when I looked up the error message there was no
prior art. This wasted about 3 hours of development time.

This commit adds a Work-In-Progress Sidebar component. It may be useful
to un-group the default `loading` and `error` views for the `channels`
DataStateView. These loading and error UIs could be defaults for the
rest of the application (may not want dependence on MUI).

This commit adds a "lessons learned" section to the README.

This commit also adds a `Color` enum export to `style.tsx` utility. This
is useful to group the theme and related color constants. It's also just
enjoyable to use.