Add react-router, Chat header, and channel routes
This commit adds the react-router-dom package, which makes it possible
to switch chat channels. One important thing to note here is that the
`useHistory` hook provided by react-router-dom must be used by a
component that is a child of a `<Router />` instance. I made that
mistake, costing about 30 minutes.

It is now possible to create a new channel in the app UI, via the
`<Sidebar />` component.

The sidebar previously did not reach the bottom of the page, it now has
`height: 100vh` like it should have from the start.

I fixed an issue with the sidebar `<Divider />` instances on Chrome.
There was an ugly issue with the border, solved by adding `border: 0;`
to remove any native styling. The issue does not exist on Firefox.

I implemented this ugly wrapper, `DbWrite`, around
`db.collection('foo').add(x)`. This is because Firebase does not allow a
type-safe mechanism for adding items to a collection. I go back and
forth pondering an ideal way to ensure items are of the right type/shape
upon adding to a collection. `DbWrite` is likely not the prettiest way,
it will likely be removed in the future in favor of type assertions.