Add ability to send and read messages
This commit refactors a core chunk of chat code. The view logic
for a chat messages lives in MessageView.tsx. This takes care of
positioning and styling. Messages do not have a clickable user image
or username yet. There are no plans for a user account page currently.

A vague bullet list is added to the README for ideas and notes while
developing.

`date-fns` package has been added for use in formatting message
timestamps. This is unnecessary but kind of cool to see the relative
time, like, "22 seconds ago."

This commit also stubs the Login page. WIP.