Add comment deletion, Login page, and sign-out
Finally fix that annoying height issue where the message input field
would render slightly below the viewport. Issue was that the height of
the non-header UI was independent of the height of the header. The
height of the app body now takes into account the height of the header
to calculate its own size. This causes the message input field AND the
sidebar height to be correctly set to 100% in their own containers. Two
birds with one stone!

The `firestore.Timestamp` value is sometimes null apparently! Attempting
to render it immediately results in accessing some field on `null`
which crashes the app (: As a result, timestamp fields on interfaces are
now `Timestamp | null`.

Add `useUser` hook to fetch the signed-in user easily from anywhere in
the app.

Add `UserImage` component for displaying user avatars with content.

Removed superfluous sidebar options.

Added green "dot" icon near user's real name.

Updated Sidebar styling to show more channels.